Extract classes to separate files
