+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/walletconnect/request/WcRequestStellarScreenPre.kt
@@ -1,0 +1,22 @@
+package io.horizontalsystems.bankwallet.modules.walletconnect.request
+
+import androidx.compose.runtime.Composable
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.navigation.NavController
+import io.horizontalsystems.bankwallet.entities.DataState
+import io.horizontalsystems.bankwallet.ui.compose.components.ListErrorView
+
+@Composable
+fun WcRequestStellarScreenPre(navController: NavController) {
+    val viewModelPre = viewModel<WCRequestStellarPreViewModel>(
+        factory = WCRequestStellarPreViewModel.Factory()
+    )
+
+    val uiState = viewModelPre.uiState
+
+    if (uiState is DataState.Success) {
+        WcRequestStellarScreen(navController, uiState.data.sessionRequest, uiState.data.wcAction)
+    } else if (uiState is DataState.Error) {
+        ListErrorView(uiState.error.message ?: "Error") { }
+    }
+}
